// src/app/api/models/mesa-electoral-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TipoMesa } from '../models/tipo-mesa';
export interface MesaElectoralDto {
  activa?: boolean;
  canton?: string | null;
  codigoMesa?: string | null;
  descripcion?: string | null;
  descripcionEleccion?: string | null;
  eleccion?: string | null;
  idMesaElectoral?: string;
  parroquia?: string | null;
  provincia?: string | null;
  tipoMesa?: TipoMesa;
  zona?: string | null;
}