// src/app/api/models/acta-eleccion.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ActaDetalle } from '../models/acta-detalle';
import { EstadoActa } from '../models/estado-acta';
import { TipoCandidato } from '../models/tipo-candidato';
export interface ActaEleccion {
  actaDetalles?: Array<ActaDetalle> | null;
  eleccionId?: string;
  estado?: EstadoActa;
  fechaModificacion?: string | null;
  fechaRegistro?: string;
  idActa?: string;
  mesaElectoralId?: string;
  tipoCandidato?: TipoCandidato;
  totalVotos?: number;
  usuarioModificacionId?: string | null;
  usuarioRegistroId?: string;
  votosBlancos?: number;
  votosNulos?: number;
}